# dmestore-ui/src/main/angular/api-doc.json
{
    "swagger": "2.0",
    "info": {
        "description": "SpingCloud web接口列表",
        "version": "1.0",
        "title": "接口文档",
        "contact": {},
        "license": {}
    },
    "host": "localhost:8080",
    "basePath": "/",
...
    "paths": {
...
        "/accessnfs/gethostgroupsbystorageid/{storageId}": {
            "get": {
                "tags": ["nfs-access-controller"],
                "summary": "getHostGroupsByStorageId",
                "operationId": "getHostGroupsByStorageIdUsingGET",
                "produces": ["*/*"],
                "parameters": [{
                    "name": "storageId",
                    "in": "path",
                    "description": "storageId",
                    "required": true,
                    "type": "string"
                }],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ResponseBodyBean"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "deprecated": false
            }
        },
        "/accessnfs/gethostsbystorageid/{storageId}": {
            "get": {
                "tags": ["nfs-access-controller"],
                "summary": "getHostsByStorageId",
                "operationId": "getHostsByStorageIdUsingGET",
                "produces": ["*/*"],
                "parameters": [{
                    "name": "storageId",
                    "in": "path",
                    "description": "storageId",
                    "required": true,
                    "type": "string"
                }],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ResponseBodyBean"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "deprecated": false
            }
        },
...
        "/accessvmfs/createvmfs": {
            "post": {
                "tags": ["vmfs-access-controller"],
                "summary": "createVmfs",
                "operationId": "createVmfsUsingPOST",
                "consumes": ["application/json"],
                "produces": ["*/*"],
                "parameters": [{
                    "in": "body",
                    "name": "params",
                    "description": "params",
                    "required": true,
                    "schema": {
                        "type": "object"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ResponseBodyBean"
                        }
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "deprecated": false
            }
        },
        "/accessvmfs/deletevmfs": {
            "post": {
                "tags": ["vmfs-access-controller"],
                "summary": "deleteVmfs",
                "operationId": "deleteVmfsUsingPOST",
                "consumes": ["application/json"],
                "produces": ["*/*"],
                "parameters": [{
                    "in": "body",
                    "name": "params",
                    "description": "params",
                    "required": true,
                    "schema": {
                        "type": "object"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ResponseBodyBean"
                        }
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "deprecated": false
            }
        },
        "/accessvmfs/gethostgroupsbystorageid/{storageId}": {
            "get": {
                "tags": ["vmfs-access-controller"],
                "summary": "getHostGroupsByStorageId",
                "operationId": "getHostGroupsByStorageIdUsingGET_1",
                "produces": ["*/*"],
                "parameters": [{
                    "name": "storageId",
                    "in": "path",
                    "description": "storageId",
                    "required": true,
                    "type": "string"
                }],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ResponseBodyBean"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "deprecated": false
            }
        },
        "/accessvmfs/gethostsbystorageid/{storageId}": {
            "get": {
                "tags": ["vmfs-access-controller"],
                "summary": "getHostsByStorageId",
                "operationId": "getHostsByStorageIdUsingGET_1",
                "produces": ["*/*"],
                "parameters": [{
                    "name": "storageId",
                    "in": "path",
                    "description": "storageId",
                    "required": true,
                    "type": "string"
                }],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ResponseBodyBean"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "deprecated": false
            }
        },
        "/accessvmfs/listvmfs": {
            "get": {
                "tags": ["vmfs-access-controller"],
                "summary": "listVmfs",
                "operationId": "listVmfsUsingGET",
                "produces": ["*/*"],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ResponseBodyBean"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "deprecated": false
            }
        },
        "/accessvmfs/listvmfsperformance": {
            "get": {
                "tags": ["vmfs-access-controller"],
                "summary": "listVmfsPerformance",
                "operationId": "listVmfsPerformanceUsingGET",
                "produces": ["*/*"],
                "parameters": [{
                    "name": "wwns",
                    "in": "query",
                    "description": "wwns",
                    "required": true,
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "collectionFormat": "multi"
                }],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ResponseBodyBean"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "deprecated": false
            }
        },
        "/accessvmfs/mountvmfs": {
            "post": {
                "tags": ["vmfs-access-controller"],
                "summary": "mountVmfs",
                "operationId": "mountVmfsUsingPOST",
                "consumes": ["application/json"],
                "produces": ["*/*"],
                "parameters": [{
                    "in": "body",
                    "name": "params",
                    "description": "params",
                    "required": true,
                    "schema": {
                        "type": "object"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ResponseBodyBean"
                        }
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "deprecated": false
            }
        },
        "/accessvmfs/querydatastorebyname": {
            "get": {
                "tags": ["vmfs-access-controller"],
                "summary": "queryDatastoreByName",
                "operationId": "queryDatastoreByNameUsingGET",
                "produces": ["*/*"],
                "parameters": [{
                    "name": "name",
                    "in": "query",
                    "description": "name",
                    "required": true,
                    "type": "string"
                }],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ResponseBodyBean"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "deprecated": false
            }
        },
        "/accessvmfs/queryvmfs": {
            "get": {
                "tags": ["vmfs-access-controller"],
                "summary": "queryVmfs",
                "operationId": "queryVmfsUsingGET",
                "produces": ["*/*"],
                "parameters": [{
                    "name": "dataStoreObjectId",
                    "in": "query",
                    "description": "dataStoreObjectId",
                    "required": true,
                    "type": "string"
                }],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ResponseBodyBean"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "deprecated": false
            }
        },
        "/accessvmfs/scanvmfs": {
            "get": {
                "tags": ["vmfs-access-controller"],
                "summary": "scanvmfs",
                "operationId": "scanvmfsUsingGET",
                "produces": ["*/*"],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ResponseBodyBean"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "deprecated": false
            }
        },
        "/accessvmfs/ummountvmfs": {
            "post": {
                "tags": ["vmfs-access-controller"],
                "summary": "unmountVmfs",
                "operationId": "unmountVmfsUsingPOST",
                "consumes": ["application/json"],
                "produces": ["*/*"],
                "parameters": [{
                    "in": "body",
                    "name": "params",
                    "description": "params",
                    "required": true,
                    "schema": {
                        "type": "object"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ResponseBodyBean"
                        }
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "deprecated": false
            }
        },
        "/accessvmfs/volume/{storageObjectId}": {
            "get": {
                "tags": ["vmfs-access-controller"],
                "summary": "volumeDetail",
                "operationId": "volumeDetailUsingGET",
                "produces": ["*/*"],
                "parameters": [{
                    "name": "storageObjectId",
                    "in": "path",
                    "description": "storageObjectId",
                    "required": true,
                    "type": "string"
                }],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ResponseBodyBean"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "deprecated": false
            }
        },
...
    },
    "definitions": {
...
        "ResponseBodyBean": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {
                    "type": "object"
                },
                "description": {
                    "type": "string"
                }
            },
            "title": "ResponseBodyBean"
        },
...
    }
}